libcamera: bayer_format: Add the transpose transformation
To transpose a BayerFormat means to flip it over its main diagonal.

For example:
G B    G R
    ->
R G    B G

The main diagonal goes from the top left to the bottom right. This
means, that the only two orders affected by a transpose are GBRG & GRBG.
When a transpose is used in combination with horizontal and/or vertical
flips it is performed after the flips.  Therefore add the functionality
by switching GBRG (index 1) with GRBG (index 2), after the flips have
been applied.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>